% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howzat-package.R
\docType{package}
\name{howzat-package}
\alias{howzat}
\alias{howzat-package}
\title{howzat: Convert Spatial Objects to 'Well-known text'}
\description{
Converts spatial objects to 'Well-known text' format  
 <https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry>.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
