% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{wkt}
\alias{wkt}
\alias{wkt.POINT}
\alias{wkt.MULTIPOINT}
\alias{wkt.LINESTRING}
\alias{wkt.MULTILINESTRING}
\alias{wkt.POLYGON}
\alias{wkt.MULTIPOLYGON}
\alias{wkt.sfc}
\alias{wkt.sf}
\title{Convert to WKT}
\usage{
wkt(x, ...)

\method{wkt}{POINT}(x, ...)

\method{wkt}{MULTIPOINT}(x, ...)

\method{wkt}{LINESTRING}(x, ...)

\method{wkt}{MULTILINESTRING}(x, ...)

\method{wkt}{POLYGON}(x, ...)

\method{wkt}{MULTIPOLYGON}(x, ...)

\method{wkt}{sfc}(x, ...)

\method{wkt}{sf}(x, ...)
}
\arguments{
\item{x}{object to format as WKT}

\item{...}{ignored}
}
\description{
Convert to Well-Known-Text geometry string format.
}
\details{
Methods exist for the simple features types.

Note, this is not the 'WKT' specification for a coordinate reference sytem
(map projection), it's a geometry encoding for points, lines, and polygons.
}
\examples{
wkt(sfobj)
wkt(sfobj$geom)
wkt(sfobj$geom[[2]])
}
